serial: fsl_linflexuart: Fix Woverflow warning on conversion
A Woverflow warning was introduced by SDK commit e11fb4e609ab ("linflex:
Correctly clear UARTSR in interrupt mode").
  drivers/tty/serial/fsl_linflexuart.c: In function ‘linflex_rxint’:
  drivers/tty/serial/fsl_linflexuart.c:629:24: warning: conversion from ‘long unsigned int’ to ‘u32’ {aka ‘unsigned int’} changes value from ‘18446744073709551613’ to ‘4294967293’ [-Woverflow]
    629 |                 writel(~LINFLEXD_UARTSR_DTFTFF, sport->membase + UARTSR)

Fix it by force casting to u32.

+++ drivers/tty/serial/fsl_linflexuart.c
@@ -626,7 +626,7 @@ static irqreturn_t linflex_rxint(int irq, void *dev_id)
 			break;
 
 		rx = readb(sport->membase + BDRM);
-		writel(~LINFLEXD_UARTSR_DTFTFF, sport->membase + UARTSR);
+		writel(~(u32)LINFLEXD_UARTSR_DTFTFF, sport->membase + UARTSR);
 
 		flg = TTY_NORMAL;
 		sport->icount.rx++;
